requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
croniter==0.3.37
freeze==3.0
graia-application-mirai==0.17.0
graia-broadcast==0.7.0
idna==2.10
iterwrapper==0.1.4
multidict==5.1.0
natsort==7.1.1
pydantic==1.7
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
